# emp-demo/src/main/java/com/example/emp_demo/repository/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/employees/{id}":{"get":{"tags":["employee-controller"],"operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"put":{"tags":["employee-controller"],"operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"delete":{"tags":["employee-controller"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/employees":{"get":{"tags":["employee-controller"],"operationId":"getAllEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Employee"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"createEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/api/employees/{id}/image":{"get":{"tags":["employee-controller"],"operationId":"downloadImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"image/png":{"schema":{"type":"string","format":"byte"}}}}}},"post":{"tags":["employee-controller"],"operationId":"uploadImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/employees/{id}/{department}":{"patch":{"tags":["employee-controller"],"operationId":"updateDepartment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"department","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/api/version-demo/v1/employee":{"get":{"tags":["versioning-demo-controller"],"operationId":"uriVersioning","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/version-demo/param":{"get":{"tags":["versioning-demo-controller"],"operationId":"paramVersioning","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/version-demo/header":{"get":{"tags":["versioning-demo-controller"],"operationId":"headerVersioning","parameters":[{"name":"X-API-VERSION","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/employees/dto-demo/{id}":{"get":{"tags":["employee-controller"],"operationId":"getEmpById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}}}},"components":{"schemas":{"Employee":{"required":["department","email","firstName","lastName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"department":{"type":"string"},"salary":{"minimum":0,"type":"number","format":"double"},"profileImageUrl":{"type":"string"}}},"EmployeeDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"department":{"type":"string"}}}}}}